Added option to skip purging mirror failures
If an error occurs during processing user / media, admin can now choose to keep
mirror for such downloads (if mirror's enabled). The reason behind this option
is to simplify debugging failed transmissions with MAL.

+++ src/Processors/AbstractProcessor.php
@@ -89,7 +89,8 @@ abstract class AbstractProcessor
 		}
 		catch (Exception $e)
 		{
-			Downloader::purgeCache($urls);
+			if (Config::$mirrorPurgeFailures)
+				Downloader::purgeCache($urls);
 			throw $e;
 		}
